frontend: make `nix-env` install option less prominent
I've seen people complain that `nix-env` is a weird "default" option to
install packages with, particularly because of its weird ergonomics:
- Install packages with `nix-env -iA nixos.hello`
- Uninstall without the `nixos` prefix

I think the community consensus has been to avoid the `nix-env` command
altogether, but at first glance, to a newcomer, since the `nix-env` tab
is currently the leftmost option (although not the one selected by
default), `nix-env` might look like the way to go...

To make the install instruction page slightly more ergonomic (and
hopefully nudge people to the correct option when they start out), I've
ordered them from left-to-right (read: reversed), and greyed out the
`nix-env` option, making it the least prominent from now on visually.

Now this might entail the fact that some screenreaders and other
structured consumers of the markup will see the options the wrong way
around, but I guess that's just a flaw of the design, where `pull-right`
is used to right-align the tabs, incidentally reversing the order....

Disclaimer: I don't have any experience with Elm so forgive me if
I struggle to follow-up further review comments or requests.